requirements.txt
tensorflow==2.1.0
numpy==1.18.3
pillow==7.1.2
flask==1.0.3